Update dependencies and targetSdkVersion
To prepare for adding more comprehensive unit tests I decided to
add mockito as a dependency. I also decided to go ahead and update
all the existing dependencies to the latest version. While I was at
it I also updated the targetSdkVersion and compileSdkVersion to 29
so that future updates don't get blocked by Google Play.